python/requirements.txt
blas==1.0
bottleneck==1.4.2
brotli-python==1.0.9
bzip2==1.0.8
ca-certificates==2025.2.25
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.11.0
fonttools==4.55.3
freetype==2.12.1
icc_rt==2022.1.0
icu==73.1
intel-openmp==2023.1.0
joblib==1.4.2
jpeg==9e
kiwisolver==1.4.8
krb5==1.20.1
lcms2==2.16
lerc==4.0.0
libdeflate==1.22
libffi==3.4.4
libpng==1.6.39
libpq==17.4
libtiff==4.5.1
libwebp-base==1.3.2
lz4-c==1.9.4
matplotlib==3.10.0
matplotlib-base==3.10.0
mkl==2023.1.0
mkl-service==2.4.0
mkl_fft==1.3.11
mkl_random==1.2.8
nltk==3.9.1
numexpr==2.10.1
openjpeg==2.5.2
openssl==3.0.16
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==25.0
pyparsing==3.2.0
pyqt==6.7.1
pyqt6-sip==13.9.1
python==3.10.16
python-dateutil==2.9.0post0
python-tzdata==2023.3
pytz==2024.1
qtbase==6.7.2
qtdeclarative==6.7.2
qtsvg==6.7.2
qttools==6.7.2
qtwebchannel==6.7.2
qtwebsockets==6.7.2
scikit-learn==1.6.1
setuptools==72.1.0
sip==6.10.0
six==1.16.0
sqlite==3.45.3
tbb==2021.8.0
threadpoolctl==3.5.0
tk==8.6.14
tomli==2.0.1
tornado==6.4.2
tqdm==4.67.1
tzdata==2025a
unicodedata2==15.1.0
vc==14.42
vs2015_runtime==14.42.34433
wheel==0.45.1
xz==5.6.4
zlib==1.2.13
zstd==1.5.6
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpathlib==0.21.0
cmdstanpy==1.2.5
comm==0.2.2
confection==0.1.5
cymem==2.0.11
cython==3.0.12
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
datasets==3.4.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
docutils==0.21.2
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flask==3.0.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gensim==4.3.3
gitdb==4.0.12
gitpython==3.1.44
h11==0.14.0
holidays==0.68
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
imagesize==1.4.1
importlib-metadata==8.6.1
importlib-resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.3.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
langcodes==3.5.0
language-data==1.3.0
lightgbm==4.6.0
llvmlite==0.44.0
logai==0.1.5
logparser==0.8.4
logparser3==1.0.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
narwhals==1.30.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.3
notebook-shim==0.2.4
numba==0.61.0
numpy==1.26.4
overrides==7.7.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
preshed==3.0.9
prometheus-client==0.21.1
prompt-toolkit==3.0.50
propcache==0.3.0
prophet==1.1.6
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py4j==0.10.9.9
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pyproject-hooks==1.2.0
pytest==8.3.5
python-json-logger==3.3.0
pywin32==310
pywinpty==2.0.15
pyyaml==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2022.3.2
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
salesforce-merlion==2.0.4
schema==0.7.7
scipy==1.13.1
seaborn==0.13.2
send2trash==1.8.3
shellingham==1.5.4
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
srsly==2.5.1
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.4
sympy==1.13.1
terminado==0.18.1
thinc==8.3.4
tinycss2==1.4.0
tokenizers==0.13.3
toml==0.10.2
torch==2.6.0
traitlets==5.14.3
transformers==4.23.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
uri-template==1.3.0
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0